2laba/backend/requirements.txt
djangorestframework
django-filter
django-cors-headers
markdown
psycopg[binary,pool]
faker
gunicorn